% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{pnadc_example}
\alias{pnadc_example}
\title{Path for example data}
\usage{
pnadc_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
Path for example data
}
\examples{
pnadc_example()
pnadc_example(path="exampledata.txt")
}
