% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadcovid19_design.R
\name{pnadcovid19_design}
\alias{pnadcovid19_design}
\title{Create pnadcovid19 survey object utilizing its sampling design for analysis with survey package}
\usage{
pnadcovid19_design(data_pnadcovid19)
}
\arguments{
\item{data_pnadcovid19}{A tibble of PNAD-COVID19 data read with \code{read_pnadcovid19} function.}
}
\value{
An object of class \code{survey.design} with the data from PNAD-COVID19 survey and its sample design.
}
\description{
Create pnadcovid19 survey object utilizing its sampling design for analysis with survey package
}
\examples{

#Using data read from disk
input_path <- pnadcovid19_example("input_example.txt")
data_path <- pnadcovid19_example("exampledata.txt")
pnadcovid19.df <- read_pnadcovid19(data_path, input_path, vars="VD4002")
dictionary.path <- pnadcovid19_example("dictionaryexample.xls")
pnadcovid19.df <- pnadcovid19_labeller(pnadcovid19.df,dictionary.path)
\dontrun{
pnadcovid19.svy <- pnadcovid19_design(pnadcovid19.df)
#Calculating unemployment rate
survey::svymean(~VD4002, pnadcovid19.svy, na.rm=TRUE)}

#Downloading data
\dontrun{
pnadcovid19.df2<- get_pnadc(2,2020,vars="VD4002")
pnadcovid19.df2 <- pnadcovid19_labeller(pnadcovid19.df2,dictionary.path)
pnadcovid19.svy2 <- pnadcovid19_design(pnadcovid19.df2)
#Calculating unemployment rate
survey::svymean(~VD4002, pnadcovid19.svy2, na.rm=TRUE)}
}
