% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadcovid19_labeller.R
\name{pnadcovid19_labeller}
\alias{pnadcovid19_labeller}
\title{Label categorical variables from PNADCovid19 datasets}
\usage{
pnadcovid19_labeller(data_pnadcovid19, dictionary.file)
}
\arguments{
\item{data_pnadcovid19}{A tibble of PNAD-covid19 data read with \code{read_pnadcovid19} function.}

\item{dictionary.file}{The dictionary file for selected survey available on official website: \url{ftp://ftp.ibge.gov.br/Trabalho_e_Rendimento/Pesquisa_Nacional_por_Amostra_de_Domicilios_PNAD_COVID19/Microdados/Documentacao/Dicionario_e_input.zip}}
}
\value{
A tibble with the data provided from PNADc survey and its categorical variables as factors with labels.
}
\description{
Label categorical variables from PNADCovid19 datasets
}
\examples{
input_path <- pnadcovid19_example("input_example.txt")
data_path <- pnadcovid19_example("exampledata.txt")
dictionary.path <- pnadcovid19_example("dictionaryexample.xls")
pnadc.df <- read_pnadcovid19(data_path, input_path, vars="VD4002")
pnadc.df <- pnadcovid19_labeller(pnadc.df,dictionary.path)

}
