% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pnadcovid19.R
\name{read_pnadcovid19}
\alias{read_pnadcovid19}
\title{Read PNAD-COVID19 microdata}
\usage{
read_pnadcovid19(microdata, input_xlsx, vars = NULL)
}
\arguments{
\item{microdata}{A text file containing microdata from PNAD-COVID19 survey. The file must be downloaded from \url{ftp://ftp.ibge.gov.br/Trabalho_e_Rendimento/Pesquisa_Nacional_por_Amostra_de_Domicilios_PNAD_COVID19/Microdados/}}

\item{vars}{Character vector of the name of the variables you want to keep for analysys. \code{default} is to keep all variables}
}
\value{
A tibble with the survey design variables and selected variables.
}
\description{
Read PNAD-COVID19 microdata
}
\examples{
input_path <- pnadcovid19_example("input_example.xlsx")
data_path <- pnadcovid19_example("exampledata.txt")
pnadc.df <- read_pnadcovid19(data_path,input_path,vars="VD4002")
}
